+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Updated check dependencies module in running script package.json for dev mode
+- Updated absolute path for component page
 
 ## 1.0.5 - October 2021